dialogue_memory_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dialogue-memory-pipeline"
version = "0.1.0"
description = "LLM-driven dialogue segmentation and episodic memory construction pipeline."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
  { name = "Jan" }
]
keywords = [
  "dialogue",
  "segmentation",
  "memory",
  "llm",
  "nlp"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "openai",
  "python-dotenv",
]

[project.optional-dependencies]
test = [
  "pytest>=8",
]

[project.urls]
Homepage = "https://github.com/Keyan0412/dialogue_memory_pipeline"
Repository = "https://github.com/Keyan0412/dialogue_memory_pipeline"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["dialogue_memory_pipeline*"]

[tool.setuptools.package-data]
dialogue_memory_pipeline = ["data/*.json"]
